+++ src/env_universal_common.cpp
@@ -40,6 +40,7 @@
 #include "env.h"
 #include "env_universal_common.h"
 #include "fallback.h"  // IWYU pragma: keep
+#include "flog.h"
 #include "path.h"
 #include "utf8.h"
 #include "util.h"  // IWYU pragma: keep
@@ -195,11 +196,11 @@ static bool append_file_entry(env_var_t::env_var_flags_t flags, const wcstring &
 
     // Append variable name like "fish_color_cwd".
     if (!valid_var_name(key_in)) {
-        debug(0, L"Illegal variable name: '%ls'", key_in.c_str());
+        FLOG(error, L"Illegal variable name: '%ls'", key_in.c_str());
         success = false;
     }
     if (success && !append_utf8(key_in, result, storage)) {
-        debug(0, L"Could not convert %ls to narrow character string", key_in.c_str());
+        FLOG(error, L"Could not convert %ls to narrow character string", key_in.c_str());
         success = false;
     }
 
@@ -210,7 +211,7 @@ static bool append_file_entry(env_var_t::env_var_flags_t flags, const wcstring &
 
     // Append value.
     if (success && !append_utf8(full_escape(val_in), result, storage)) {
-        debug(0, L"Could not convert %ls to narrow character string", val_in.c_str());
+        FLOG(error, L"Could not convert %ls to narrow character string", val_in.c_str());
         success = false;
     }
 
@@ -447,7 +448,8 @@ bool env_universal_t::write_to_fd(int fd, const wcstring &path) {
     std::string contents = serialize_with_vars(vars);
     if (write_loop(fd, contents.data(), contents.size()) < 0) {
         const char *error = std::strerror(errno);
-        debug(0, _(L"Unable to write to universal variables file '%ls': %s"), path.c_str(), error);
+        FLOG(error, _(L"Unable to write to universal variables file '%ls': %s"), path.c_str(),
+             error);
         success = false;
     }
 
@@ -462,8 +464,8 @@ bool env_universal_t::move_new_vars_file_into_place(const wcstring &src, const w
     int ret = wrename(src, dst);
     if (ret != 0) {
         const char *error = std::strerror(errno);
-        debug(0, _(L"Unable to rename file from '%ls' to '%ls': %s"), src.c_str(), dst.c_str(),
-              error);
+        FLOG(error, _(L"Unable to rename file from '%ls' to '%ls': %s"), src.c_str(), dst.c_str(),
+             error);
     }
     return ret == 0;
 }
@@ -525,7 +527,7 @@ bool env_universal_t::open_temporary_file(const wcstring &directory, wcstring *o
 
     if (!success) {
         const char *error = std::strerror(saved_errno);
-        debug(0, _(L"Unable to open temporary file '%ls': %s"), out_path->c_str(), error);
+        FLOG(error, _(L"Unable to open temporary file '%ls': %s"), out_path->c_str(), error);
     }
     return success;
 }
@@ -587,7 +589,8 @@ bool env_universal_t::open_and_acquire_lock(const wcstring &path, int *out_fd) {
             }
 #endif
             const char *error = std::strerror(errno);
-            debug(0, _(L"Unable to open universal variable file '%ls': %s"), path.c_str(), error);
+            FLOG(error, _(L"Unable to open universal variable file '%ls': %s"), path.c_str(),
+                 error);
             break;
         }
 
@@ -1058,7 +1061,7 @@ class universal_notifier_shmem_poller_t : public universal_notifier_t {
         int fd = shm_open(path, O_RDWR | O_CREAT, 0600);
         if (fd < 0) {
             const char *error = std::strerror(errno);
-            debug(0, _(L"Unable to open shared memory with path '%s': %s"), path, error);
+            FLOG(error, _(L"Unable to open shared memory with path '%s': %s"), path, error);
             errored = true;
         }
 
@@ -1068,8 +1071,8 @@ class universal_notifier_shmem_poller_t : public universal_notifier_t {
             struct stat buf = {};
             if (fstat(fd, &buf) < 0) {
                 const char *error = std::strerror(errno);
-                debug(0, _(L"Unable to fstat shared memory object with path '%s': %s"), path,
-                      error);
+                FLOG(error, _(L"Unable to fstat shared memory object with path '%s': %s"), path,
+                     error);
                 errored = true;
             }
             size = buf.st_size;
@@ -1079,7 +1082,8 @@ class universal_notifier_shmem_poller_t : public universal_notifier_t {
         bool set_size = !errored && size < (off_t)sizeof(universal_notifier_shmem_t);
         if (set_size && ftruncate(fd, sizeof(universal_notifier_shmem_t)) < 0) {
             const char *error = std::strerror(errno);
-            debug(0, _(L"Unable to truncate shared memory object with path '%s': %s"), path, error);
+            FLOG(error, _(L"Unable to truncate shared memory object with path '%s': %s"), path,
+                 error);
             errored = true;
         }
 
@@ -1089,8 +1093,8 @@ class universal_notifier_shmem_poller_t : public universal_notifier_t {
                               MAP_SHARED, fd, 0);
             if (addr == MAP_FAILED) {
                 const char *error = std::strerror(errno);
-                debug(0, _(L"Unable to memory map shared memory object with path '%s': %s"), path,
-                      error);
+                FLOG(error, _(L"Unable to memory map shared memory object with path '%s': %s"),
+                     path, error);
                 this->region = NULL;
             } else {
                 this->region = static_cast<universal_notifier_shmem_t *>(addr);
@@ -1495,7 +1499,7 @@ void universal_notifier_named_pipe_t::make_pipe(const wchar_t *test_path) {
     if (mkfifo_status == -1 && errno != EEXIST) {
         const char *error = std::strerror(errno);
         const wchar_t *errmsg = _(L"Unable to make a pipe for universal variables using '%ls': %s");
-        debug(0, errmsg, vars_path.c_str(), error);
+        FLOG(error, errmsg, vars_path.c_str(), error);
         pipe_fd = -1;
         return;
     }
@@ -1504,7 +1508,7 @@ void universal_notifier_named_pipe_t::make_pipe(const wchar_t *test_path) {
     if (fd < 0) {
         const char *error = std::strerror(errno);
         const wchar_t *errmsg = _(L"Unable to open a pipe for universal variables using '%ls': %s");
-        debug(0, errmsg, vars_path.c_str(), error);
+        FLOG(error, errmsg, vars_path.c_str(), error);
         pipe_fd = -1;
         return;
     }